Create README.md
How to Use the Telegram Bot
Start a chat with the bot
Find the bot at: @BonusHABot

Send commands
Use the available commands to interact with the bot. You can send commands like /models to get a list of available models, or send a command to train a model or make predictions.